Assistant checkpoint: Fix gas price parsing for decimal values
Assistant generated file changes:
- main.py: Fix gas price parsing to handle decimal values
Fix gas price parsing to handle decimal values
Fix BSC gas price parsing to handle decimal values
Fix gas price parsing to handle decimal values
Fix BSC gas price parsing to handle decimal values
Fix Polygon gas price parsing to handle decimal values
Fix gas price parsing to handle decimal values
Fix BSC gas price parsing to handle decimal values
Fix Polygon gas price parsing to handle decimal values

---

User prompt:

при тестировании бота происходит ❌ Ошибка получения данных Ethereum и ❌ Ошибка получения данных BSC

+++ main.py
@@ -1,4 +1,3 @@
-
 import asyncio
 import logging
 import os
@@ -24,15 +23,15 @@ class BlockchainFeesBot:
         self.bot_token = os.getenv('TELEGRAM_BOT_TOKEN')
         if not self.bot_token:
             raise ValueError("TELEGRAM_BOT_TOKEN не найден в переменных окружения")
-        
+
         self.application = Application.builder().token(self.bot_token).build()
         self.setup_handlers()
-    
+
     def setup_handlers(self):
         """Настройка обработчиков команд и коллбэков"""
         self.application.add_handler(CommandHandler("start", self.start_command))
         self.application.add_handler(CallbackQueryHandler(self.button_callback))
-    
+
     async def start_command(self, update: Update, context: ContextTypes.DEFAULT_TYPE):
         """Обработчик команды /start"""
         keyboard = [
@@ -53,23 +52,23 @@ class BlockchainFeesBot:
                 InlineKeyboardButton("🔷 Arbitrum", callback_data="arbitrum")
             ]
         ]
-        
+
         reply_markup = InlineKeyboardMarkup(keyboard)
-        
+
         welcome_text = (
             "🤖 Добро пожаловать в бот мониторинга комиссий блокчейнов!\n\n"
             "Выберите блокчейн для просмотра текущих комиссий:"
         )
-        
+
         await update.message.reply_text(welcome_text, reply_markup=reply_markup)
-    
+
     async def button_callback(self, update: Update, context: ContextTypes.DEFAULT_TYPE):
         """Обработчик нажатий на инлайн-кнопки"""
         query = update.callback_query
         await query.answer()
-        
+
         blockchain = query.data
-        
+
         try:
             fees_info = await self.get_blockchain_fees(blockchain)
             await query.edit_message_text(text=fees_info)
@@ -78,7 +77,7 @@ class BlockchainFeesBot:
             await query.edit_message_text(
                 text=f"❌ Ошибка получения данных для {blockchain.upper()}. Попробуйте позже."
             )
-    
+
     async def get_token_price(self, token_id: str) -> float:
         """Получение цены токена через CoinGecko API"""
         try:
@@ -116,7 +115,7 @@ class BlockchainFeesBot:
         except Exception as e:
             logger.error(f"Ошибка получения данных для {blockchain}: {e}")
             raise
-    
+
     async def get_ethereum_fees(self) -> str:
         """Получение комиссий Ethereum через Etherscan API"""
         try:
@@ -125,23 +124,23 @@ class BlockchainFeesBot:
                 timeout=10
             )
             data = response.json()
-            
+
             if data['status'] == '1':
-                safe = int(data['result']['SafeGasPrice'])
-                standard = int(data['result']['ProposeGasPrice'])
-                fast = int(data['result']['FastGasPrice'])
-                
+                safe = float(data['result']['SafeGasPrice'])
+                standard = float(data['result']['ProposeGasPrice'])
+                fast = float(data['result']['FastGasPrice'])
+
                 # Получаем цену ETH
                 eth_price = await self.get_token_price("ethereum")
-                
+
                 # Стандартная транзакция ETH ~21000 gas
                 gas_limit = 21000
-                
+
                 if eth_price:
                     safe_usd = (safe * gas_limit * 0.000000001) * eth_price
                     standard_usd = (standard * gas_limit * 0.000000001) * eth_price
                     fast_usd = (fast * gas_limit * 0.000000001) * eth_price
-                    
+
                     return (
                         f"🔵 **Ethereum (ETH)**\n\n"
                         f"⚡ Быстрая: {fast} Gwei (≈ ${fast_usd:.3f})\n"
@@ -163,7 +162,7 @@ class BlockchainFeesBot:
         except Exception as e:
             logger.error(f"Ошибка API Ethereum: {e}")
             return "❌ Ошибка получения данных Ethereum"
-    
+
     async def get_bsc_fees(self) -> str:
         """Получение комиссий BSC через BscScan API"""
         try:
@@ -172,23 +171,23 @@ class BlockchainFeesBot:
                 timeout=10
             )
             data = response.json()
-            
+
             if data['status'] == '1':
-                safe = int(data['result']['SafeGasPrice'])
-                standard = int(data['result']['ProposeGasPrice'])
-                fast = int(data['result']['FastGasPrice'])
-                
+                safe = float(data['result']['SafeGasPrice'])
+                standard = float(data['result']['ProposeGasPrice'])
+                fast = float(data['result']['FastGasPrice'])
+
                 # Получаем цену BNB
                 bnb_price = await self.get_token_price("binancecoin")
-                
+
                 # Стандартная транзакция BSC ~21000 gas
                 gas_limit = 21000
-                
+
                 if bnb_price:
                     safe_usd = (safe * gas_limit * 0.000000001) * bnb_price
                     standard_usd = (standard * gas_limit * 0.000000001) * bnb_price
                     fast_usd = (fast * gas_limit * 0.000000001) * bnb_price
-                    
+
                     return (
                         f"🟡 **BSC (BNB)**\n\n"
                         f"⚡ Быстрая: {fast} Gwei (≈ ${fast_usd:.4f})\n"
@@ -210,7 +209,7 @@ class BlockchainFeesBot:
         except Exception as e:
             logger.error(f"Ошибка API BSC: {e}")
             return "❌ Ошибка получения данных BSC"
-    
+
     async def get_bitcoin_fees(self) -> str:
         """Получение комиссий Bitcoin через Mempool.space API"""
         try:
@@ -219,26 +218,26 @@ class BlockchainFeesBot:
                 timeout=10
             )
             data = response.json()
-            
+
             fast = data['fastestFee']
             half_hour = data['halfHourFee']
             hour = data['hourFee']
-            
+
             # Получаем цену BTC
             btc_price = await self.get_token_price("bitcoin")
-            
+
             # Средняя транзакция Bitcoin ~250 байт
             tx_size = 250
-            
+
             if btc_price:
                 fast_btc = fast * tx_size * 0.00000001
                 half_hour_btc = half_hour * tx_size * 0.00000001
                 hour_btc = hour * tx_size * 0.00000001
-                
+
                 fast_usd = fast_btc * btc_price
                 half_hour_usd = half_hour_btc * btc_price
                 hour_usd = hour_btc * btc_price
-                
+
                 return (
                     f"🟠 **Bitcoin (BTC)**\n\n"
                     f"⚡ Быстрая (~10 мин): {fast} sat/vB (≈ ${fast_usd:.2f})\n"
@@ -258,18 +257,18 @@ class BlockchainFeesBot:
         except Exception as e:
             logger.error(f"Ошибка API Bitcoin: {e}")
             return "❌ Ошибка получения данных Bitcoin"
-    
+
     async def get_solana_fees(self) -> str:
         """Получение информации о комиссиях Solana"""
         try:
             # Получаем курс SOL
             sol_price = await self.get_token_price("solana")
-            
+
             fee_sol = 0.000005
-            
+
             if sol_price:
                 fee_usd = fee_sol * sol_price
-                
+
                 return (
                     f"🟢 **Solana (SOL)**\n\n"
                     f"💰 Стандартная комиссия: {fee_sol:.6f} SOL (≈ ${fee_usd:.6f})\n"
@@ -289,22 +288,22 @@ class BlockchainFeesBot:
                 f"💰 Стандартная комиссия: 0.000005 SOL\n\n"
                 f"💡 Фиксированная комиссия для большинства транзакций"
             )
-    
+
     async def get_ton_fees(self) -> str:
         """Получение комиссий TON через Tonapi"""
         try:
             # Получаем цену TON
             ton_price = await self.get_token_price("the-open-network")
-            
+
             fee_low = 0.005
             fee_standard = 0.01
             fee_high = 0.02
-            
+
             if ton_price:
                 fee_low_usd = fee_low * ton_price
                 fee_standard_usd = fee_standard * ton_price
                 fee_high_usd = fee_high * ton_price
-                
+
                 return (
                     f"🟣 **TON**\n\n"
                     f"💰 Простая транзакция: ~{fee_low} TON (≈ ${fee_low_usd:.4f})\n"
@@ -329,20 +328,20 @@ class BlockchainFeesBot:
                 f"📊 Комиссия зависит от сложности транзакции\n\n"
                 f"💡 Очень низкие комиссии за счет архитектуры"
             )
-    
+
     async def get_tron_fees(self) -> str:
         """Получение информации о комиссиях Tron"""
         try:
             # Получаем курс TRX
             trx_price = await self.get_token_price("tron")
-            
+
             bandwidth_fee = 0.001
             energy_fee = 15
-            
+
             if trx_price:
                 bandwidth_usd = bandwidth_fee * trx_price
                 energy_usd = energy_fee * trx_price
-                
+
                 return (
                     f"🔴 **Tron (TRX)**\n\n"
                     f"📡 Обычный перевод: {bandwidth_fee} TRX (≈ ${bandwidth_usd:.6f})\n"
@@ -365,7 +364,7 @@ class BlockchainFeesBot:
                 f"⚡ Смарт-контракт: ~15 TRX\n\n"
                 f"💡 Обычные переводы: очень дешево"
             )
-    
+
     async def get_polygon_fees(self) -> str:
         """Получение комиссий Polygon через PolygonScan API"""
         try:
@@ -374,23 +373,23 @@ class BlockchainFeesBot:
                 timeout=10
             )
             data = response.json()
-            
+
             if data['status'] == '1':
-                safe = int(data['result']['SafeGasPrice'])
-                standard = int(data['result']['ProposeGasPrice'])
-                fast = int(data['result']['FastGasPrice'])
-                
+                safe = float(data['result']['SafeGasPrice'])
+                standard = float(data['result']['ProposeGasPrice'])
+                fast = float(data['result']['FastGasPrice'])
+
                 # Получаем цену MATIC
                 matic_price = await self.get_token_price("matic-network")
-                
+
                 # Стандартная транзакция Polygon ~21000 gas
                 gas_limit = 21000
-                
+
                 if matic_price:
                     safe_usd = (safe * gas_limit * 0.000000001) * matic_price
                     standard_usd = (standard * gas_limit * 0.000000001) * matic_price
                     fast_usd = (fast * gas_limit * 0.000000001) * matic_price
-                    
+
                     return (
                         f"🟪 **Polygon (MATIC)**\n\n"
                         f"⚡ Быстрая: {fast} Gwei (≈ ${fast_usd:.5f})\n"
@@ -417,7 +416,7 @@ class BlockchainFeesBot:
                 f"📊 Комиссия зависит от загрузки сети\n\n"
                 f"💡 Намного дешевле Ethereum"
             )
-    
+
     async def get_arbitrum_fees(self) -> str:
         """Получение комиссий Arbitrum через Arbiscan API"""
         try:
@@ -425,26 +424,26 @@ class BlockchainFeesBot:
                 "https://api.arbiscan.io/api?module=gastracker&action=gasoracle",
                 timeout=10
             )
-            
+
             if response.status_code == 200:
                 data = response.json()
-                
+
                 if data.get('status') == '1' and 'result' in data:
                     safe = int(data['result']['SafeGasPrice'])
                     standard = int(data['result']['ProposeGasPrice'])
                     fast = int(data['result']['FastGasPrice'])
-                    
+
                     # Получаем цену ETH
                     eth_price = await self.get_token_price("ethereum")
-                    
+
                     # Стандартная транзакция ~21000 gas
                     gas_limit = 21000
-                    
+
                     if eth_price:
                         safe_usd = (safe * gas_limit * 0.000000001) * eth_price
                         standard_usd = (standard * gas_limit * 0.000000001) * eth_price
                         fast_usd = (fast * gas_limit * 0.000000001) * eth_price
-                        
+
                         return (
                             f"🔷 **Arbitrum (ETH)**\n\n"
                             f"⚡ Быстрая: {fast} Gwei (≈ ${fast_usd:.5f})\n"
@@ -465,24 +464,24 @@ class BlockchainFeesBot:
                     logger.warning(f"Неожиданный ответ API Arbitrum: {data}")
             else:
                 logger.warning(f"HTTP {response.status_code} от API Arbitrum")
-                
+
         except requests.RequestException as e:
             logger.error(f"Ошибка запроса к API Arbitrum: {e}")
         except Exception as e:
             logger.error(f"Неожиданная ошибка API Arbitrum: {e}")
-            
+
         # Возвращаем фиксированную информацию при любой ошибке
         eth_price = await self.get_token_price("ethereum")
-        
+
         if eth_price:
             # Типичные комиссии Arbitrum
             low_gwei = 0.1
             high_gwei = 2.0
             gas_limit = 21000
-            
+
             low_usd = (low_gwei * gas_limit * 0.000000001) * eth_price
             high_usd = (high_gwei * gas_limit * 0.000000001) * eth_price
-            
+
             return (
                 f"🔷 **Arbitrum (ETH)**\n\n"
                 f"💰 Типичная комиссия: 0.1-2 Gwei (≈ ${low_usd:.5f}-${high_usd:.4f})\n"
@@ -501,7 +500,7 @@ class BlockchainFeesBot:
                 f"💡 Layer 2 решение для Ethereum\n"
                 f"🔄 Данные API временно недоступны"
             )
-    
+
     def run(self):
         """Запуск бота"""
         logger.info("Запуск Telegram бота...")
@@ -510,7 +509,7 @@ class BlockchainFeesBot:
 def main():
     # Запускаем keep_alive в отдельном потоке
     threading.Thread(target=keep_alive, daemon=True).start()
-    
+
     # Создаем и запускаем бота
     bot = BlockchainFeesBot()
     bot.run()
